Removed stupid shit
