+++ src/commands/index.ts
@@ -5,7 +5,6 @@ import compilers from "./compilers";
 import config from "./config";
 import debug from "./debug";
 import github from "./github";
-import help from "./help";
 import permissions from "./permissions";
 import ping from "./ping";
 import report from "./report";
@@ -21,7 +20,6 @@ const commands = {
 	config,
 	debug,
 	github,
-	help,
 	permissions,
 	ping,
 	report,
